Unblock relationship

+++ user/src/modules/relationships/services/unblock-relationship.ts
@@ -1,4 +1,4 @@
-import { RelationshipType } from "@flickswipe/common";
+import { BadRequestError, RelationshipType } from "@flickswipe/common";
 import { natsWrapper } from "../../../nats-wrapper";
 import { RelationshipUnblockedPublisher } from "../events/publishers/relationship-unblocked";
 import { Relationship } from "../models/relationship";
@@ -6,31 +6,31 @@ import { Relationship } from "../models/relationship";
 export async function unblockRelationship(
   fromUserId: string,
   toUserId: string
-): Promise<boolean> {
-  // Delete current relationship
-  try {
-    await Relationship.deleteOne({
-      relationshipType: RelationshipType.Blocked,
-      sourceUser: fromUserId,
-      targetUser: toUserId,
-    });
-
-    // remove opposite relationship if not "blocked"
-    await Relationship.deleteOne({
-      relationshipType: { $ne: RelationshipType.Blocked },
-      sourceUser: toUserId,
-      targetUser: fromUserId,
-    });
-  } catch (err) {
-    console.log(`Couldn't unblock relationship`, err);
-    return false;
-  }
-
-  // publish event
-  new RelationshipUnblockedPublisher(natsWrapper.client).publish({
-    sourceUserId: fromUserId,
-    targetUserId: toUserId,
+): Promise<void> {
+  // check if current relationship exists
+  const existingRelationshipDoc = await Relationship.findOne({
+    relationshipType: RelationshipType.Blocked,
+    sourceUser: fromUserId,
+    targetUser: toUserId,
   });
 
-  return true;
+  if (!existingRelationshipDoc) {
+    throw new BadRequestError(`User is not blocked`);
+  }
+
+  // delete relationship
+  await Relationship.deleteOne({
+    relationshipType: RelationshipType.Blocked,
+    sourceUser: fromUserId,
+    targetUser: toUserId,
+  });
+
+  // publish event
+  await new RelationshipUnblockedPublisher(natsWrapper.client).publish({
+    sourceUserId: fromUserId,
+    targetUserId: toUserId,
+    updatedAt: new Date(),
+  });
+
+  console.log(`${fromUserId} unblocked relationship with ${toUserId}`);
 }
